CPR-141 move PNCIdentifier into model package
