+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/jpa/converter/PNCIdentifierConverter.kt
@@ -2,7 +2,7 @@ package uk.gov.justice.digital.hmpps.personrecord.jpa.converter
 
 import jakarta.persistence.AttributeConverter
 import jakarta.persistence.Converter
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 
 @Converter(autoApply = true)
 class PNCIdentifierConverter : AttributeConverter<PNCIdentifier, String> {
