+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/jpa/entity/OffenderEntity.kt
@@ -13,8 +13,8 @@ import jakarta.persistence.Table
 import org.hibernate.envers.Audited
 import uk.gov.justice.digital.hmpps.personrecord.client.model.OffenderDetail
 import uk.gov.justice.digital.hmpps.personrecord.jpa.converter.PNCIdentifierConverter
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
 import java.time.LocalDate
 
 @Entity
