+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/jpa/repository/DefendantRepository.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.personrecord.jpa.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.DefendantEntity
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 
 @Repository
 interface DefendantRepository : JpaRepository<DefendantEntity, Long> {
