+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/matcher/DefendantMatcher.kt
@@ -1,8 +1,8 @@
 package uk.gov.justice.digital.hmpps.personrecord.service.matcher
 
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.DefendantEntity
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
 
 class DefendantMatcher(defendants: List<DefendantEntity>?, person: Person) :
   Matcher<DefendantEntity>(defendants, person) {
