+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/matcher/OffenderMatcher.kt
@@ -1,8 +1,8 @@
 package uk.gov.justice.digital.hmpps.personrecord.service.matcher
 
 import uk.gov.justice.digital.hmpps.personrecord.client.model.OffenderDetail
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
 
 class OffenderMatcher(offenderDetails: List<OffenderDetail>?, person: Person) :
   Matcher<OffenderDetail>(offenderDetails, person) {
