+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/OffenderServiceTest.kt
@@ -17,12 +17,12 @@ import uk.gov.justice.digital.hmpps.personrecord.client.model.SearchDto
 import uk.gov.justice.digital.hmpps.personrecord.config.FeatureFlag
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.model.OtherIdentifiers
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
 import uk.gov.justice.digital.hmpps.personrecord.service.OffenderService
 import uk.gov.justice.digital.hmpps.personrecord.service.PersonRecordService
 import uk.gov.justice.digital.hmpps.personrecord.service.TelemetryService
 import uk.gov.justice.digital.hmpps.personrecord.service.type.TelemetryEventType
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
 import java.time.LocalDate
 
 @Suppress("INLINE_FROM_HIGHER_PLATFORM")
