+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/PrisonerServiceTest.kt
@@ -17,9 +17,9 @@ import uk.gov.justice.digital.hmpps.personrecord.client.model.PrisonerMatchCrite
 import uk.gov.justice.digital.hmpps.personrecord.config.FeatureFlag
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.model.OtherIdentifiers
+import uk.gov.justice.digital.hmpps.personrecord.model.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.Person
 import uk.gov.justice.digital.hmpps.personrecord.service.type.TelemetryEventType
-import uk.gov.justice.digital.hmpps.personrecord.validate.PNCIdentifier
 import java.time.LocalDate
 
 @ExtendWith(MockitoExtension::class)
